requirements.txt
requests
beautifulsoup4
selenium
pandas
python-dotenv
lxml